fix(project): update Hexalith.Security packages and add tests
Updated the version of the Hexalith.Security packages from 2.1.5 to 2.1.6 in the Directory.Packages.props file. Added a new project, HexalithApp.UnitTests, to the Hexalith.Documents.sln solution file and updated the solution configuration to include build configurations for this new project.

Additionally, modified the DocumentSummaryGrid.razor file to replace the previous download and link icons with a new component, DocumentDownloadAnchors, which handles document downloads and link creation.